refactor(models): densify fat tests and add ≤400 test-size gate
Split world_market/game/app_events/diplomacy_models suites into
concern files under the 400-line ceiling, share minimalGame fixtures,
and register repo.models_test_file_size (Refs #4068 Slice D).
