+++ SPEC/program/repo-lint.md
@@ -152,6 +152,7 @@
 | `tool/check_combat_test_scenario_harness.dart` | Forbid bare `for (final scenario in …)` loops under `packages/colonizethis_combat/test/**/*_test.dart`; scenario tables must register via `runLabeledScenarios` / `runLabeledScenarioGroup` from `colonizethis_combat_test_support` (Refs #3983); rule `repo.combat_test_scenario_harness` |
 | `tool/check_economy_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_economy/lib/**` so DealMatcher matching/indexing sub-files stay proper libraries with explicit imports (mirrors world/turn/diplomacy/orders no-part gates; Refs #3979); rule `repo.economy_no_part_directives` |
 | `tool/check_models_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_models/lib/**` so orders/world_state/world_market/app_events sub-files stay proper libraries with explicit imports (mirrors economy/world/turn no-part gates; Refs #4068); rule `repo.models_no_part_directives` |
+| `tool/check_models_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_models/test/**` exceeds 400 physical lines (aligns with diplomacy/economy test file caps; Refs #4068 Slice D); rule `repo.models_test_file_size` |
 | `tool/check_world_topology_identity_home.dart` | Require topology identity helpers (`indexTopologyNodesByRegion`, `provinceTopologyNodeId`, `provinceIdsAdjacentToSeaZone`, `regionIdForSeaZone`) to be defined in `packages/colonizethis_world/lib/src/world/topology_identity.dart` and forbid re-declaring them as top-level functions in `naval.dart` (re-export allowed; Refs #3968); rule `repo.world_topology_identity_home` |
 | `tool/check_world_test_support_builders_used.dart` | Require named world test-support builders (`spyRevealFogGame`, `ownershipTransferVisibilityGame`, `coastalSeaVisibilityGame`, `capitalLossGame`, `gpCapitalLossGame`, `minorCapitalLossGame`, `gpCapitalReassignmentGame`, `factionCapitalReassignmentGame`, `ordersPhaseGame`, `topologyGraph`) to have ≥1 call site under `packages/colonizethis_world/test/**` outside `world_test_support/` so scenario factories cannot rot unused (Refs #3968 / #4038); rule `repo.world_test_support_builders_used` |
 | `tool/check_world_fog_connectivity_no_owner_map_rebuild.dart` | Forbid fog/connectivity world lib files from calling `ownerByProvinceIdMap(...)`; ownership must use `ProvinceOwnerCache.of` (Refs #3978); rule `repo.world_fog_connectivity_no_owner_map_rebuild` — see `SPEC/program/world-fog-connectivity-ownership-sot.md` |
@@ -490,6 +491,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_economy/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #3979).
 - Given the `colonizethis_models` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4068).
 - Given a `packages/colonizethis_models/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4068).
+- Given every `.dart` file under `packages/colonizethis_models/test/**` is at or below 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the rule passes without violations and exits `0` (Refs #4068).
+- Given a `packages/colonizethis_models/test/**` Dart file with more than 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_test_file_size`, then the run fails naming that file and exits `1` (Refs #4068).
 - Given `topology_identity.dart` defines the four identity helpers and `naval.dart` does not re-declare them as top-level functions, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_topology_identity_home`, then the rule passes and exits `0` (Refs #3968).
 - Given `naval.dart` contains a top-level function declaration named `provinceTopologyNodeId`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_topology_identity_home`, then the run fails listing the offending `path:line` and exits `1` (Refs #3968).
 - Given each required world support builder (`spyRevealFogGame`, `ownershipTransferVisibilityGame`, `coastalSeaVisibilityGame`, `capitalLossGame`, `gpCapitalLossGame`, `minorCapitalLossGame`, `gpCapitalReassignmentGame`, `factionCapitalReassignmentGame`, `ordersPhaseGame`, `topologyGraph`) has ≥1 call site under `packages/colonizethis_world/test/**` outside `world_test_support/`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_test_support_builders_used`, then the rule passes and exits `0` (Refs #3968 / #4038).
